+++ sdk/resourcemanager/authorization/armauthorization/zz_generated_accessreviewscheduledefinitionsassignedformyapproval_client.go
@@ -11,11 +11,12 @@ package armauthorization
 import (
 	"context"
 	"fmt"
-	"net/http"
-
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
+	armruntime "github.com/Azure/azure-sdk-for-go/sdk/azcore/arm/runtime"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
+	"net/http"
 )
 
 // AccessReviewScheduleDefinitionsAssignedForMyApprovalClient contains the methods for the AccessReviewScheduleDefinitionsAssignedForMyApproval group.
@@ -26,8 +27,15 @@ type AccessReviewScheduleDefinitionsAssignedForMyApprovalClient struct {
 }
 
 // NewAccessReviewScheduleDefinitionsAssignedForMyApprovalClient creates a new instance of AccessReviewScheduleDefinitionsAssignedForMyApprovalClient with the specified values.
-func NewAccessReviewScheduleDefinitionsAssignedForMyApprovalClient(con *arm.Connection) *AccessReviewScheduleDefinitionsAssignedForMyApprovalClient {
-	return &AccessReviewScheduleDefinitionsAssignedForMyApprovalClient{ep: con.Endpoint(), pl: con.NewPipeline(module, version)}
+func NewAccessReviewScheduleDefinitionsAssignedForMyApprovalClient(credential azcore.TokenCredential, options *arm.ClientOptions) *AccessReviewScheduleDefinitionsAssignedForMyApprovalClient {
+	cp := arm.ClientOptions{}
+	if options != nil {
+		cp = *options
+	}
+	if len(cp.Host) == 0 {
+		cp.Host = arm.AzurePublicCloud
+	}
+	return &AccessReviewScheduleDefinitionsAssignedForMyApprovalClient{ep: string(cp.Host), pl: armruntime.NewPipeline(module, version, credential, &cp)}
 }
 
 // List - Get access review instances assigned for my approval.