+++ sdk/resourcemanager/authorization/armauthorization/zz_generated_eligiblechildresources_client.go
@@ -12,12 +12,13 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"net/http"
-	"strings"
-
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
+	armruntime "github.com/Azure/azure-sdk-for-go/sdk/azcore/arm/runtime"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
+	"net/http"
+	"strings"
 )
 
 // EligibleChildResourcesClient contains the methods for the EligibleChildResources group.
@@ -28,8 +29,15 @@ type EligibleChildResourcesClient struct {
 }
 
 // NewEligibleChildResourcesClient creates a new instance of EligibleChildResourcesClient with the specified values.
-func NewEligibleChildResourcesClient(con *arm.Connection) *EligibleChildResourcesClient {
-	return &EligibleChildResourcesClient{ep: con.Endpoint(), pl: con.NewPipeline(module, version)}
+func NewEligibleChildResourcesClient(credential azcore.TokenCredential, options *arm.ClientOptions) *EligibleChildResourcesClient {
+	cp := arm.ClientOptions{}
+	if options != nil {
+		cp = *options
+	}
+	if len(cp.Host) == 0 {
+		cp.Host = arm.AzurePublicCloud
+	}
+	return &EligibleChildResourcesClient{ep: string(cp.Host), pl: armruntime.NewPipeline(module, version, credential, &cp)}
 }
 
 // Get - Get the child resources of a resource on which user has eligible access