+++ sdk/resourcemanager/authorization/armauthorization/zz_generated_pagers.go
@@ -10,11 +10,10 @@ package armauthorization
 
 import (
 	"context"
-	"net/http"
-	"reflect"
-
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
+	"net/http"
+	"reflect"
 )
 
 // AccessReviewInstanceDecisionsListPager provides operations for iterating over paged responses.