+++ sdk/resourcemanager/authorization/armauthorization/zz_generated_provideroperationsmetadata_client.go
@@ -12,12 +12,13 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"net/http"
-	"strings"
-
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
+	armruntime "github.com/Azure/azure-sdk-for-go/sdk/azcore/arm/runtime"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
+	"net/http"
+	"strings"
 )
 
 // ProviderOperationsMetadataClient contains the methods for the ProviderOperationsMetadata group.
@@ -28,8 +29,15 @@ type ProviderOperationsMetadataClient struct {
 }
 
 // NewProviderOperationsMetadataClient creates a new instance of ProviderOperationsMetadataClient with the specified values.
-func NewProviderOperationsMetadataClient(con *arm.Connection) *ProviderOperationsMetadataClient {
-	return &ProviderOperationsMetadataClient{ep: con.Endpoint(), pl: con.NewPipeline(module, version)}
+func NewProviderOperationsMetadataClient(credential azcore.TokenCredential, options *arm.ClientOptions) *ProviderOperationsMetadataClient {
+	cp := arm.ClientOptions{}
+	if options != nil {
+		cp = *options
+	}
+	if len(cp.Host) == 0 {
+		cp.Host = arm.AzurePublicCloud
+	}
+	return &ProviderOperationsMetadataClient{ep: string(cp.Host), pl: armruntime.NewPipeline(module, version, credential, &cp)}
 }
 
 // Get - Gets provider operations metadata for the specified resource provider.