+++ sdk/resourcemanager/authorization/armauthorization/zz_generated_rolemanagementpolicies_client.go
@@ -12,13 +12,14 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
+	armruntime "github.com/Azure/azure-sdk-for-go/sdk/azcore/arm/runtime"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
 	"net/http"
 	"net/url"
 	"strings"
-
-	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
-	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
-	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
 )
 
 // RoleManagementPoliciesClient contains the methods for the RoleManagementPolicies group.
@@ -29,8 +30,15 @@ type RoleManagementPoliciesClient struct {
 }
 
 // NewRoleManagementPoliciesClient creates a new instance of RoleManagementPoliciesClient with the specified values.
-func NewRoleManagementPoliciesClient(con *arm.Connection) *RoleManagementPoliciesClient {
-	return &RoleManagementPoliciesClient{ep: con.Endpoint(), pl: con.NewPipeline(module, version)}
+func NewRoleManagementPoliciesClient(credential azcore.TokenCredential, options *arm.ClientOptions) *RoleManagementPoliciesClient {
+	cp := arm.ClientOptions{}
+	if options != nil {
+		cp = *options
+	}
+	if len(cp.Host) == 0 {
+		cp.Host = arm.AzurePublicCloud
+	}
+	return &RoleManagementPoliciesClient{ep: string(cp.Host), pl: armruntime.NewPipeline(module, version, credential, &cp)}
 }
 
 // Delete - Delete a role management policy